rename util to holidayUtils and directory common to utils
